+++ pyproject.toml
@@ -101,4 +101,4 @@ Issues = "https://github.com/huggingface/lighteval/issues"
 # Changelog = "https://github.com/huggingface/lighteval/blob/master/CHANGELOG.md"
 
 [project.scripts]
-lighteval = "lighteval.commands.lighteval_cli:main"
+lighteval = "lighteval.__main__:cli_evaluate"
